Refactor kube, ocp clients out of base and into main.  This will aid unit tests, add more badges to README
