+++ exporters/committime/collector_bitbucket.py
@@ -3,8 +3,8 @@ from collector_base import AbstractCommitCollector
 
 class BitbucketCommitCollector(AbstractCommitCollector):
 
-    def __init__(self, username, token, namespaces, apps):
-        super().__init__(username, token, namespaces, apps, "BitBucket", "TODO-DateFormat")
+    def __init__(self, kube_client, username, token, namespaces, apps):
+        super().__init__(kube_client, username, token, namespaces, apps, "BitBucket", "TODO-DateFormat")
 
     def get_commit_time(self, metric):
         """Method called to collect data and send to Prometheus"""
