+++ exporters/committime/collector_github.py
@@ -10,8 +10,8 @@ class GitHubCommitCollector(AbstractCommitCollector):
     _prefix = _prefix_pattern % _defaultapi
     _suffix = "/commits/"
 
-    def __init__(self, username, token, namespaces, apps, git_api=None):
-        super().__init__(username, token, namespaces, apps, "GitHub", '%Y-%m-%dT%H:%M:%SZ', git_api)
+    def __init__(self, kube_client, username, token, namespaces, apps, git_api=None):
+        super().__init__(kube_client, username, token, namespaces, apps, "GitHub", '%Y-%m-%dT%H:%M:%SZ', git_api)
         if git_api is not None and len(git_api) > 0:
             logging.info("Using non-default API: %s" % (git_api))
         else:
